Crate pxfm

Functions

cbrtf

Cbrt for given value for const context. This is simplified version just to make a good approximation on const context.

ceil

ceilf

copysignfk

Copies sign from y to x

copysignk

Copies sign from y to x

dirty_powf

Dirty fast pow

exp

Exp for given value for const context. This is simplified version just to make a good approximation on const context.

expf

Exp for given value for const context. This is simplified version just to make a good approximation on const context.

f_acos

Computes acos(x)

f_acosf

Compute acos

f_acosh

Huperbolic acos

f_acoshf

Hyperbolic arc cosine function

f_acospi

Computes acos(x)/PI

f_acospif

Computes acos(x)/PI

f_asin

Computes asin(x)

f_asinf

Computes asin

f_asinh

Huperbolic sine function

f_asinhf

Hyperbolic arcsine function

f_asinpi

Computes asin(x)/PI

f_asinpif

Computes asin(x)/PI

f_atan

Computes atan in double precision

f_atan2

Computes atan(x)

f_atan2f

Computes atan2

f_atan2pi

Computes atan(x)/PI

f_atan2pif

Computes atan(x/y) / PI

f_atanf

Computes atan

f_atanh

Hyperbolic arc tangent

f_atanhf

Hyperbolic atan

f_atanpi

Computes atan(x)/pi

f_atanpif

Computes atan(x)/PI

f_beta

Computes beta function

f_betaf

Computes beta function

f_betainc_reg

Regularized incomplete beta

f_betainc_regf

Regularized incomplete beta

f_cathetus

Computes the missing leg of a right triangle

f_cathetusf

Computes the missing leg of a right triangle

f_cbrt

Computes cube root

f_cbrtf

Computes cube root

f_compound

Computes (1+x)^y

f_compound_m1

Computes (1+x)^y - 1

f_compound_m1f

Computes compound (1.0 + x)^y - 1

f_compoundf

Computes compound function (1.0 + x)^y

f_cos

Cosine for double precision

f_cosf

Computes cosine function

f_cosh

Hyperbolic cosine function

f_coshf

Hyperbolic cos

f_cosm1

Computes cos(x) - 1

f_cosm1f

Computes cos(x) - 1

f_cospi

Computes cos(PI*x)

f_cospif

Computes cos(PI*x)

f_cot

Cotangent in double precision

f_cotf

Computes cotangent

f_cotpi

Computes cotangent 1/tan(PI*x)

f_cotpif

Computes 1/tan(PI*x)

f_csc

Cosecant for double precision

f_cscf

Cosecant ( 1 / sin(x) )

f_digamma

Computes digamma(x)

f_digammaf

Computes digamma(x)

f_erf

Error function

f_erfc

Complementary error function

f_erfcf

Complementary error function

f_erfcinv

Complementary inverse error function

f_erfcinvf

Complementary inverse error function

f_erfcx

Scaled complementary error function (exp(x^2)*erfc(x))

f_erfcxf

Scaled complementary error function (exp(x^2)*erfc(x))

f_erff

Error function

f_erfinv

Inverse error function

f_erfinvf

Inverse error function

f_exp

Computes exponent

f_exp2

Computes exp2

f_exp2f

Computing exp2f

f_exp2m1

Computes 2^x - 1

f_exp2m1f

Computes 2^x - 1

f_exp10

Computes exp10

f_exp10f

Computes exp10

f_exp10m1

Computes 10^x - 1

f_exp10m1f

Computes 10^x - 1

f_expf

Computes exp

f_expm1

Computes e^x - 1

f_expm1f

Computes e^x - 1

f_gamma_p

Regularized lower incomplete gamma

f_gamma_pf

Regularized lower incomplete gamma

f_gamma_q

Regularized upper incomplete gamma

f_gamma_qf

Regularized upper incomplete gamma

f_hypot

Computes hypot

f_hypot3f

f_hypotf

Hypot function

f_i0

Modified Bessel of the first kind of order 0

f_i0e

Modified exponentially scaled Bessel of the first kind of order 0

f_i0ef

Modified exponentially scaled Bessel of the first kind of order 0

f_i0f

Modified Bessel of the first kind of order 0

f_i1

Modified Bessel of the first kind of order 1

f_i2

Modified bessel of the first kind of order 2

f_i1e

Modified exponentially scaled Bessel of the first kind of order 1

f_i1ef

Modified exponentially scaled Bessel of the first kind of order 1

f_i1f

Modified Bessel of the first kind of order 1

f_i2f

Modified Bessel of the first kind of order 2

f_j0

Bessel of the first kind of order 0

f_j0f

Bessel of the first kind of order 0

f_j1

Bessel of the first kind of order 1

f_j1f

Bessel of the first kind of order 1

f_jincpi

Normalized jinc 2*J1(PI*x)/(pi*x)

f_jincpif

Normalized jinc 2*J1(PI*x)/(pi*x)

f_k0

Modified Bessel of the second kind of order 0

f_k0e

Modified exponentially scaled Bessel of the first kind of order 0

f_k0ef

Modified exponentially scaled Bessel of the first kind of order 0

f_k0f

Modified Bessel of the second kind of order 0

f_k1

Modified Bessel of the second kind of order 1

f_k1e

Modified exponentially scaled Bessel of the second kind of order 1

f_k1ef

Modified exponentially scaled Bessel of the second kind of order 1

f_k1f

Modified Bessel of the second kind of order 1

f_k2f

Modified Bessel of the second kind of order 2

f_lgamma

Computes log(gamma(x))

f_lgamma_r

Computes log(gamma(x))

f_lgamma_rf

Computes log(gamma(x))

f_lgammaf

Computes log(gamma(x))

f_lnbeta

Computes log(beta(x)) function

f_lnbetaf

Computes log(beta(x)) function

f_log

Natural logarithm

f_log2

Log2(x)

f_log1p

Computes log(x+1)

f_log1pf

Computes log(x+1)

f_log1pmx

Computes log(1+x) - x

f_log1pmxf

Computes log(1+x) - x

f_log2f

Logarithm of base 2

f_log2p1

Computes log2(x+1)

f_log2p1f

Computes log2(x+1)

f_log10

Logarithm of base 10

f_log10f

Logarithm of base 10

f_log10p1

Computes log10(x+1)

f_log10p1f

Computes log10(x+1)

f_logf

Natural logarithm

f_logistic

Logistic function

f_logisticf

Logistic function

f_logit

Inverse logistic function

f_logitf

Inverse logistic function

f_pow

Power function

f_powf

Power function

f_powm1

Computes x^y - 1

f_powm1f

Computes x^y - 1

f_rcbrt

Computes 1/cbrt(x)

f_rcbrtf

Computes 1/cbrt(x)

f_rerf

Computes 1/erf(x)

f_rerff

Computes 1/erf(x)

f_rsqrt

Computes 1/sqrt(x)

f_rsqrtf

Computes 1/sqrt(x)

f_sec

Secant for double precision

f_secf

Computes secant ( 1 / cos(x) )

f_sin

Sine for double precision

f_sinc

Computes sinc(x)

f_sincf

Computes sinc(x)

f_sincos

Sine and cosine for double precision

f_sincosf

Sine and cosine

f_sincospi

Computes sin(PIx) and cos(PIx)

f_sincospif

Computes sin(x) and cos(x) at the same time

f_sincpi

Computes sin(PI*x)/(PI*x)

f_sincpif

Computes sin(PI*x)/(PI*x)

f_sinf

Sine function

f_sinh

Hyperbolic sine function

f_sinhf

Huperbolic sine function

f_sinmx

Computes sin(x) - x

f_sinmxf

Computes sin(x) - x

f_sinpi

Computes sin(PI*x)

f_sinpif

Computes sin(PI*x)

f_sqrt1pm1

Computes sqrt(1+x) - 1

f_sqrt1pm1f

Computes sqrt(1+x) - 1

f_tan

Tangent in double precision

f_tanf

Computes tan

f_tanh

Hyperbolic tan

f_tanhf

Hyperbolic tangent

f_tanpi

Computes tan(PI*x)

f_tanpif

Computes tan(PI*x)

f_tgamma

Computes gamma(x)

f_tgammaf

True gamma function

f_trigamma

Computes the trigamma function ψ₁(x).

f_trigammaf

Computes the trigamma function ψ₁(x).

f_y0

Bessel of the second kind of order 0 (Y0)

f_y0f

Bessel of the second kind of order 0 (Y0)

f_y1

Bessel of the second kind of order 1 ( Y1 )

f_y1f

Bessel of the second kind of order 1 (Y1)

floor

Floors value

floorf

Round to integer towards minus infinity

log

Log for given value for const context. This is simplified version just to make a good approximation on const context.

logf

Log for given value for const context. This is simplified version just to make a good approximation on const context.

pow

Pow for given value for const context. This is simplified version just to make a good approximation on const context.

powf

Power function for given value for const context. This is simplified version just to make a good approximation on const context.

round

round_ties_even

roundf

roundf_ties_even

sqrtf

Computes Square root. Most of CPU have built-in instruction with higher precision, prefer use this only for const contexts.

trunc

truncf