pxfm::bessel

Module j0

Functions

f_j0

Bessel of the first kind of order 0

j0_asympt 🔒

j0_asympt_fast 🔒

j0_asympt_hard 🔒

j0_maclaurin_series 🔒

Generated by SageMath:

j0_maclaurin_series_fast 🔒

Generated by SageMath:

j0_maclaurin_series_hard 🔒

Generated by SageMath:

j0_small_argument_dd 🔒

j0_small_argument_fast 🔒

This method on small range searches for nearest zero or extremum. Then picks stored series expansion at the point end evaluates the poly at the point.

j0_small_argument_hard 🔒