pxfm

Module common

Constants

EXP_MASK_F32 🔒

Functions

copysignfk

Copies sign from y to x

copysignk

Copies sign from y to x

dd_fmla 🔒

Executes mandatory FMA if not available will be simulated through Dekker and Veltkamp

dd_fmlaf 🔒

dyad_fmla 🔒

f_fmla 🔒

Optional FMA, if it is available hardware FMA will use, if not then just scalar c + a * b

f_fmlaf 🔒

fmla 🔒

fmlaf 🔒

is_integer 🔒

is_integerf 🔒

is_odd_integer 🔒

is_odd_integerf 🔒

mask_trailing_ones_u32 🔒

min_normal_f64 🔒

pow2i 🔒

Computes 2^n

pow2if 🔒

rintfk 🔒

rintk 🔒

Round towards whole integral number

set_exponent_f32 🔒