pxfm::tangent

Module atan2pi

Constants

MPI 🔒

MPI_OVER_2 🔒

MZERO 🔒

PI 🔒

PI_OVER_2 🔒

PI_OVER_4 🔒

THREE_PI_OVER_4 🔒

ZERO 🔒

Statics

CONST_ADJ 🔒

IS_NEG 🔒

Functions

atan2pi_big_exp_difference_hard 🔒

If one of arguments is too huge or too small, extended precision is required for case with big exponent difference

atan2pi_fma_impl 🔒 ⚠

atan2pi_gen_impl 🔒

f_atan2pi

Computes atan(x)/PI